requirements.txt
Django==3.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.1
ordered-set==4.0.2
psycopg2-binary
PyJWT
python-dateutil
pytz==2021.1
six
sqlparse
typing-extensions